# docx/deliverables/openapi.yaml
openapi: 3.1.0
info:
  title: 学生生活一站式社区 AI 助手 API
  version: 0.5.0
  description: |
    详细设计增量 1：公共基础与 M4 平台治理；增量 2：M2 校园服务中心；
    增量 3：M1 AI 知识与对话；增量 4：M3 校园社区与互助；
    增量 5：M5 多智能体、Tool Runtime、小模型工程、训练评估与 M2 演示电费。
    已发布 operationId 不得改变。
  contact:
    name: 项目开发小组
  license:
    name: Educational Use Only
    identifier: LicenseRef-Educational-Use-Only
servers:
  - url: http://localhost:8000
    description: 后端本地开发
  - url: http://localhost
    description: Docker Compose 演示环境
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/community/anonymous-identities/reveal:
    post:
      tags: [CommunityInteractions]
      operationId: revealAnonymousIdentity
      summary: 按明确事由反查匿名内容作者
      description: 不进入普通运营菜单；每次成功或失败均写 community.anonymous_identity.read 审计。
      x-permissions: [community:anonymous_identity:read]
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AnonymousIdentityRevealRequest"}
      responses:
        "200":
          description: 匿名内容作者身份；响应禁止缓存
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AnonymousIdentityRevealResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /api/v1/service-guides:
    get:
      tags: [ServiceGuides]
      operationId: listServiceGuides
      summary: 查询已发布且有效的办事指南
      parameters:
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: q
          in: query
          description: 按标题和摘要搜索
          schema: {type: string, maxLength: 100}
        - name: category_code
          in: query
          schema: {type: string, maxLength: 50}
        - name: department_id
          in: query
          schema: {type: string, format: uuid}
        - name: campus_code
          in: query
          schema: {type: string, maxLength: 30}
        - name: student_type
          in: query
          schema: {$ref: "#/components/schemas/StudentType"}
      responses:
        "200":
          description: 指南分页结果
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ServiceGuidePageResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/service-guides/{guide_id}:
    get:
      tags: [ServiceGuides]
      operationId: getServiceGuide
      summary: 获取办事指南详情
      parameters:
        - $ref: "#/components/parameters/GuideId"
        - $ref: "#/components/parameters/XRequestId"
        - name: campus_code
          in: query
          required: true
          schema: {type: string, maxLength: 30}
        - name: student_type
          in: query
          required: true
          schema: {$ref: "#/components/schemas/StudentType"}
      responses:
        "200":
          description: 包含适用对象、材料、步骤、地点、时间和联系人
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ServiceGuideResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/service-guides/{guide_id}/checklist:
    get:
      tags: [ServiceGuides]
      operationId: getServiceGuideChecklist
      summary: 按校区和学生类型生成可解释材料清单
      parameters:
        - $ref: "#/components/parameters/GuideId"
        - $ref: "#/components/parameters/XRequestId"
        - name: campus_code
          in: query
          required: true
          schema: {type: string, maxLength: 30}
        - name: student_type
          in: query
          required: true
          schema: {$ref: "#/components/schemas/StudentType"}
      responses:
        "200":
          description: 与适用条件匹配的材料清单及命中原因
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MaterialChecklistResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /api/v1/electricity/accounts/{room_id}:
    get:
      tags: [Electricity]
      operationId: getElectricityBalance
      summary: 查询本人绑定房间的演示电费余额
      parameters:
        - $ref: "#/components/parameters/RoomId"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: 查询成功；source 固定为 mock
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ElectricityBalanceResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/electricity/topup-requests:
    post:
      tags: [Electricity]
      operationId: createElectricityTopupRequest
      summary: 创建模拟电费充值申请
      description: 仅改变演示数据；不接真实支付。直接从 UI 调用时由用户点击确认，Agent 调用时还需 approval_id。
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ElectricityTopupRequest"}
      responses:
        "201":
          description: 模拟充值申请已创建
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ElectricityTopupResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    internalServiceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
      description: 仅受信 Agent Runtime 使用的独立服务凭证，不接受用户 JWT
...
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: 客户端可提供；非法或缺失时服务端生成新 ID
      schema: {type: string, minLength: 8, maxLength: 64}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 同一用户和接口 24 小时内重放返回首次结果
      schema: {type: string, minLength: 8, maxLength: 128}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    GuideId:
      name: guide_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    RoomId:
      name: room_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    Unauthorized:
      description: Access Token 缺失、无效或已过期
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          examples:
            unauthenticated:
              $ref: "#/components/examples/UnauthorizedError"
    Forbidden:
      description: 已认证但缺少操作或资源权限
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          example:
            code: AUTH_FORBIDDEN
            message: 没有执行该操作的权限
            details: []
            request_id: 01J2ABCDEF0123456789ABCDE0
            timestamp: 2026-07-12T08:00:00Z
    NotFound:
      description: 资源不存在，或当前用户无权感知该资源
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Conflict:
      description: 唯一约束、状态机或乐观锁冲突
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          example:
            code: RESOURCE_VERSION_CONFLICT
            message: 数据已被其他操作更新，请刷新后重试
            details: []
            request_id: 01J2ABCDEF0123456789ABCDE0
            timestamp: 2026-07-12T08:00:00Z
    ValidationError:
      description: 字段格式或业务输入校验失败
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
          example:
            code: VALIDATION_ERROR
            message: 请求参数校验失败
            details:
              - field: password
                reason: 至少 10 个字符，并包含字母、数字和特殊字符
            request_id: 01J2ABCDEF0123456789ABCDE0
            timestamp: 2026-07-12T08:00:00Z
...
  examples:
    UnauthorizedError:
      value:
        code: AUTH_UNAUTHORIZED
        message: 登录状态无效，请重新登录
        details: []
        request_id: 01J2ABCDEF0123456789ABCDE0
        timestamp: 2026-07-12T08:00:00Z
  schemas:
    ResponseMeta:
      type: object
      required: [code, message, request_id, timestamp]
      properties:
        code: {type: string, example: OK}
        message: {type: string, example: success}
        request_id: {type: string, maxLength: 64}
        timestamp: {type: string, format: date-time}
...
    ErrorDetail:
      type: object
      required: [reason]
      additionalProperties: false
      properties:
        field: {type: [string, "null"]}
        reason: {type: string}
        context:
          type: [object, "null"]
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [code, message, details, request_id, timestamp]
      additionalProperties: false
      properties:
        code: {type: string}
        message: {type: string}
        details:
          type: array
          items: {$ref: "#/components/schemas/ErrorDetail"}
        request_id: {type: string}
        timestamp: {type: string, format: date-time}
    PageMeta:
      type: object
      required: [page, page_size, total, total_pages]
      additionalProperties: false
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
        total_pages: {type: integer, minimum: 0}
...
    AnonymousIdentityRevealRequest:
      type: object
      required: [target_type, target_id, reason]
      additionalProperties: false
      properties:
        target_type: {type: string, enum: [post, comment]}
        target_id: {type: string, format: uuid}
        reason: {type: string, minLength: 2, maxLength: 500}
    AnonymousIdentityReveal:
      type: object
      required: [target_type, target_id, author_user_id, username, display_name, reason, revealed_at]
      additionalProperties: false
      properties:
        target_type: {type: string, enum: [post, comment]}
        target_id: {type: string, format: uuid}
        author_user_id: {type: string, format: uuid}
        username: {type: string}
        display_name: {type: string}
        reason: {type: string}
        revealed_at: {type: string, format: date-time}
    AnonymousIdentityRevealResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/AnonymousIdentityReveal"}
...
    StudentType:
      type: string
      enum: [undergraduate, postgraduate, international, all]
    DepartmentContact:
      type: object
      required: [id, department_id, campus_code, office_name, location, valid_from]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        department_id: {type: string, format: uuid}
        campus_code: {type: string}
        contact_name: {type: [string, "null"]}
        office_name: {type: string}
        phone: {type: [string, "null"]}
        email: {type: [string, "null"], format: email}
        location: {type: string}
        office_hours: {type: [string, "null"]}
        valid_from: {type: string, format: date}
        valid_until: {type: [string, "null"], format: date}
    DepartmentSummary:
      type: object
      required: [id, code, name]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        description: {type: [string, "null"]}
...
    GuideCategory:
      type: object
      required: [code, name]
      additionalProperties: false
      properties:
        code: {type: string}
        name: {type: string}
    GuideApplicability:
      type: object
      required: [campus_code, student_type, applicable]
      additionalProperties: false
      properties:
        campus_code: {type: string}
        student_type: {$ref: "#/components/schemas/StudentType"}
        applicable: {type: boolean}
        notes: {type: [string, "null"]}
    GuideMaterial:
      type: object
      required: [id, name, required, copies, included, inclusion_reason, sort_order]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        description: {type: [string, "null"]}
        required: {type: boolean}
        copies: {type: integer, minimum: 0, maximum: 20}
        included: {type: boolean}
        inclusion_reason: {type: string, description: 说明通用规则或命中的校区/学生类型条件}
        sort_order: {type: integer, minimum: 0}
    GuideStep:
      type: object
      required: [step_no, title, description]
      additionalProperties: false
      properties:
        step_no: {type: integer, minimum: 1}
        title: {type: string}
        description: {type: string}
        location: {type: [string, "null"]}
        estimated_minutes: {type: [integer, "null"], minimum: 1}
    ServiceGuideSummary:
      type: object
      required: [id, code, title, summary, category, department, updated_at, version]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        title: {type: string}
        summary: {type: string}
        category: {$ref: "#/components/schemas/GuideCategory"}
        department: {$ref: "#/components/schemas/DepartmentSummary"}
        location: {type: [string, "null"]}
        service_hours: {type: [string, "null"]}
        valid_until: {type: [string, "null"], format: date}
        updated_at: {type: string, format: date-time}
        version: {type: integer, minimum: 1}
    ServiceGuideDetail:
      type: object
      required: [id, code, title, summary, category, department, applicability, materials, steps, contacts, updated_at, version]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        title: {type: string}
        summary: {type: string}
        category: {$ref: "#/components/schemas/GuideCategory"}
        department: {$ref: "#/components/schemas/DepartmentSummary"}
        location: {type: [string, "null"]}
        service_hours: {type: [string, "null"]}
        source_url: {type: [string, "null"], format: uri}
        valid_until: {type: [string, "null"], format: date}
        applicability: {$ref: "#/components/schemas/GuideApplicability"}
        materials:
          type: array
          items: {$ref: "#/components/schemas/GuideMaterial"}
        steps:
          type: array
          items: {$ref: "#/components/schemas/GuideStep"}
        contacts:
          type: array
          items: {$ref: "#/components/schemas/DepartmentContact"}
        updated_at: {type: string, format: date-time}
        version: {type: integer, minimum: 1}
    ServiceGuidePageData:
      type: object
      required: [items, pagination]
      additionalProperties: false
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/ServiceGuideSummary"}
        pagination: {$ref: "#/components/schemas/PageMeta"}
    ServiceGuidePageResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/ServiceGuidePageData"}
    ServiceGuideResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/ServiceGuideDetail"}
    MaterialChecklistData:
      type: object
      required: [guide_id, campus_code, student_type, applicable, materials]
      additionalProperties: false
      properties:
        guide_id: {type: string, format: uuid}
        campus_code: {type: string}
        student_type: {$ref: "#/components/schemas/StudentType"}
        applicable: {type: boolean}
        applicability_reason: {type: [string, "null"]}
        materials:
          type: array
          items: {$ref: "#/components/schemas/GuideMaterial"}
    MaterialChecklistResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/MaterialChecklistData"}
...
    ElectricityBalance:
      type: object
      required: [room_id, room_name, balance_cny, source, is_simulated, as_of]
      additionalProperties: false
      properties:
        room_id: {type: string, format: uuid}
        room_name: {type: string, maxLength: 100}
        balance_cny: {type: number, minimum: 0, multipleOf: 0.01}
        source: {type: string, const: mock}
        is_simulated: {type: boolean, const: true}
        as_of: {type: string, format: date-time}
    ElectricityBalanceResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/ElectricityBalance"}
    ElectricityTopupRequest:
      type: object
      required: [room_id, amount_cny]
      additionalProperties: false
      properties:
        room_id: {type: string, format: uuid}
        amount_cny: {type: number, minimum: 1, maximum: 500, multipleOf: 0.01}
        approval_id:
          description: Agent 发起时必填；UI 直接发起时为空
          type: [string, "null"]
          format: uuid
        agent_run_id: {type: [string, "null"], format: uuid}
    ElectricityTopup:
      type: object
      required: [request_id, room_id, amount_cny, status, source, is_simulated, notice, created_at]
      additionalProperties: false
      properties:
        request_id: {type: string, format: uuid}
        room_id: {type: string, format: uuid}
        amount_cny: {type: number, multipleOf: 0.01}
        status: {type: string, const: simulated}
        source: {type: string, const: mock}
        is_simulated: {type: boolean, const: true}
        notice: {type: string, maxLength: 300}
        created_at: {type: string, format: date-time}
    ElectricityTopupResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required: [data]
          properties:
            data: {$ref: "#/components/schemas/ElectricityTopup"}
...